+++ Assets/_Game/Scripts/PlayerScripts/SOWeaponData.cs
@@ -10,6 +10,6 @@ public class SOWeaponData : ScriptableObject
     public float bulletSpeed = 2f;
     public float fireRate = 0.5f;
     public int damage = 20;
-    public int resourceAmount;
+    public float resourceAmount;
     public AudioClip fireClip;
 }
